Move CodeStarSource class to its own module
